backend/requirements.txt
asgiref
certifi
charset-normalizer
cloudinary==1.32.0
dj-database-url
Django
django-ckeditor
django-cloudinary-storage
django-cors-headers
django-environ
django-js-asset
djangorestframework
gunicorn
idna
psycopg2
Pillow
requests
pytz
six
sqlparse
typing_extensions
urllib3==1.26.12
whitenoise
